Separates out modifiers from DateRangePicker class and into a DayPickerWithModifiers class
